feat: add create rgb invoice
The test cases in invoice/mod.rs fail on the Windows environment. However, they generate a warning similar to the Debian 12 environment. Despite the warning, invoices can still be generated. It is likely that the recipient of the subsequent contract will need to use the -force option to accept the contract.